gcp/requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
joblib>=1.3.0

# Google Cloud Platform
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.11.0
google-cloud-aiplatform>=1.35.0
google-cloud-functions>=1.8.0
google-cloud-run>=0.10.0
google-cloud-logging>=3.8.0
google-cloud-monitoring>=2.15.0

# Data processing
apache-beam[gcp]>=2.46.0
pyarrow>=12.0.0

# API and web
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0

# Containerization
docker>=6.1.0
kubernetes>=26.1.0

# Infrastructure
terraform>=1.5.0 